Move ContentEdit out
- Move ContentEdit ADT to its own file
- Rename match variables to reflect parameter names
- Validate position & length
